Fix NodePort issues with SRV records and Node labels
This commit fixes two main issues with the external-dns implementation
of NodePort based DNS records:

(1) Previously, the serviceSource would consider all nodes a target,
regardless of their k8s role. This resulted in the master's public IP
being added to the DNS record. Given that most k8s users do not use the
master for running services/pods and reserve it for k8s API activity,
this does not seem like the ideal default behavior. Instead, this commit
gives the user the ability to specify what "role" label a node should
have, then only selects nodes based on that role. This defaults to the
"node" role, which by default refers to all non-master (i.e., worker)
nodes in the cluster.

(2) The earlier implementation of NodePort DNS created SRV records
alongside the other DNS records created for NodePort. However, this
functionality was broken by the introduction of a subsequent commit:
https://github.com/kubernetes-incubator/external-dns/commit/d934f693fffb224fc1f8852722a95b8b731c25e6#r32415167
which updated the Planner's filterRecordsForPlan to only recognize
CNAMEs and A records. This commit adds SRVs to the recognized records.

However, since no other service types get SRV records created for them
and SRV support is not well documented in external-dns, this commit
toggles the creation of SRV records using a flag which is disabled by
default.

Finally, since this commit introduces two new flags to the
serviceSource, it also takes the liberty of cleaning up the
NewServiceSource() factory to take in the config struct instead of an unruly
list of flags and settings. There are a number of other long
factory functions that may benefit from a similar refactor.